Move ProxyAuthScheme to SPI package
Per team review, this makes more sense in the SPI package where it can
be shared by other client implementations.
